requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
chromadb==0.5.3
sentence-transformers==3.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
streamlit==1.38.0
ollama==0.3.3
numpy==1.26.4